+++ services/horizon/internal/app.go
@@ -532,6 +532,8 @@ func (a *App) init() error {
 		SSEUpdateFrequency:       a.config.SSEUpdateFrequency,
 		StaleThreshold:           a.config.StaleThreshold,
 		ConnectionTimeout:        a.config.ConnectionTimeout,
+		ClientQueryTimeout:       a.config.ClientQueryTimeout,
+		MaxConcurrentRequests:    a.config.MaxConcurrentRequests,
 		MaxHTTPRequestSize:       a.config.MaxHTTPRequestSize,
 		NetworkPassphrase:        a.config.NetworkPassphrase,
 		MaxPathLength:            a.config.MaxPathLength,
@@ -552,6 +554,7 @@ func (a *App) init() error {
 			},
 			cache: newHealthCache(healthCacheTTL),
 		},
+		SkipTxMeta: a.config.SkipTxmeta,
 	}
 
 	if a.primaryHistoryQ != nil {