+++ services/horizon/internal/flags.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"os/exec"
 	"strings"
+	"time"
 
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
@@ -68,7 +69,9 @@ const (
 	// StellarTestnet is a constant representing the Stellar test network
 	StellarTestnet = "testnet"
 
-	defaultMaxHTTPRequestSize = uint(200 * 1024)
+	defaultMaxConcurrentRequests = uint(1000)
+	defaultMaxHTTPRequestSize    = uint(200 * 1024)
+	clientQueryTimeoutNotSet     = -1
 )
 
 var (
@@ -438,11 +441,44 @@ func Flags() (*Config, support.ConfigOptions) {
 			UsedInCommands: ApiServerCommands,
 		},
 		&support.ConfigOption{
+			Name:        "client-query-timeout",
+			ConfigKey:   &config.ClientQueryTimeout,
+			OptType:     types.Int,
+			FlagDefault: clientQueryTimeoutNotSet,
+			CustomSetValue: func(co *support.ConfigOption) error {
+				if !support.IsExplicitlySet(co) {
+					*(co.ConfigKey.(*time.Duration)) = time.Duration(co.FlagDefault.(int))
+					return nil
+				}
+				duration := viper.GetInt(co.Name)
+				if duration < 0 {
+					return fmt.Errorf("%s cannot be negative", co.Name)
+				}
+				*(co.ConfigKey.(*time.Duration)) = time.Duration(duration) * time.Second
+				return nil
+			},
+			Usage: "defines the timeout for when horizon will cancel all postgres queries connected to an HTTP request. The timeout is measured in seconds since the start of the HTTP request. Note, this timeout does not apply to POST /transactions. " +
+				"The difference between client-query-timeout and connection-timeout is that connection-timeout applies a postgres statement timeout whereas client-query-timeout will send an additional request to postgres to cancel the ongoing query. " +
+				"Generally, client-query-timeout should be configured to be higher than connection-timeout to allow the postgres statement timeout to kill long running queries without having to send the additional cancel request to postgres. " +
+				"By default, client-query-timeout will be set to twice the connection-timeout. Setting client-query-timeout to 0 will disable the timeout which means that Horizon will never kill long running queries using the cancel request, however, " +
+				"long running queries can still be killed through the postgres statement timeout which is configured via the connection-timeout flag.",
+			UsedInCommands: ApiServerCommands,
+		},
+		&support.ConfigOption{
 			Name:           "max-http-request-size",
 			ConfigKey:      &config.MaxHTTPRequestSize,
 			OptType:        types.Uint,
 			FlagDefault:    defaultMaxHTTPRequestSize,
 			Usage:          "sets the limit on the maximum allowed http request payload size, default is 200kb, to disable the limit check, set to 0, only do so if you acknowledge the implications of accepting unbounded http request payload sizes.",
+			UsedInCommands: ApiServerCommands,
+		},
+		&support.ConfigOption{
+			Name:        "max-concurrent-requests",
+			ConfigKey:   &config.MaxConcurrentRequests,
+			OptType:     types.Uint,
+			FlagDefault: defaultMaxConcurrentRequests,
+			Usage: "sets the limit on the maximum number of concurrent http requests, default is 1000, to disable the limit set to 0. " +
+				"If Horizon receives a request which would exceed the limit of concurrent http requests, Horizon will respond with a 503 status code.",
 			UsedInCommands: ApiServerCommands,
 		},
 		&support.ConfigOption{
@@ -981,6 +1017,11 @@ func ApplyFlags(config *Config, flags support.ConfigOptions, options ApplyOption
 	if config.BehindCloudflare && config.BehindAWSLoadBalancer {
 		return fmt.Errorf("invalid config: Only one option of --behind-cloudflare and --behind-aws-load-balancer is allowed." +
 			" If Horizon is behind both, use --behind-cloudflare only")
+	}
+
+	if config.ClientQueryTimeout == clientQueryTimeoutNotSet {
+		// the default value for cancel-db-query-timeout is twice the connection-timeout
+		config.ClientQueryTimeout = config.ConnectionTimeout * 2
 	}
 
 	return nil