+++ services/horizon/internal/integration/transaction_test.go
@@ -63,13 +63,7 @@ func TestP19MetaDisabledTransaction(t *testing.T) {
 
 	clientTx := itest.MustSubmitOperations(&masterAccount, itest.Master(), op)
 
-	var txMetaResult xdr.TransactionMeta
-	err = xdr.SafeUnmarshalBase64(clientTx.ResultMetaXdr, &txMetaResult)
-	require.NoError(t, err)
-
-	assert.Equal(t, len(txMetaResult.MustV2().Operations), 0)
-	assert.Equal(t, len(txMetaResult.MustV2().TxChangesAfter), 0)
-	assert.Equal(t, len(txMetaResult.MustV2().TxChangesBefore), 0)
+	assert.Empty(t, clientTx.ResultMetaXdr)
 }
 
 func TestP20MetaTransaction(t *testing.T) {
@@ -123,12 +117,5 @@ func TestP20MetaDisabledTransaction(t *testing.T) {
 	preFlightOp, minFee := itest.PreflightHostFunctions(&sourceAccount, *installContractOp)
 	clientTx := itest.MustSubmitOperationsWithFee(&sourceAccount, itest.Master(), minFee+txnbuild.MinBaseFee, &preFlightOp)
 
-	var txMetaResult xdr.TransactionMeta
-	err = xdr.SafeUnmarshalBase64(clientTx.ResultMetaXdr, &txMetaResult)
-	require.NoError(t, err)
-
-	assert.Equal(t, len(txMetaResult.MustV3().Operations), 0)
-	assert.Nil(t, txMetaResult.MustV3().SorobanMeta)
-	assert.Equal(t, len(txMetaResult.MustV3().TxChangesAfter), 0)
-	assert.Equal(t, len(txMetaResult.MustV3().TxChangesBefore), 0)
+	assert.Empty(t, clientTx.ResultMetaXdr)
 }